+++ src/style/style_layer.ts
@@ -128,7 +128,7 @@ export abstract class StyleLayer extends Evented {
         this.type = layer.type;
         this._globalState = globalState;
         this._featureFilter = {filter: () => true, needGeometry: false, getGlobalStateRefs: () => new Set<string>()};
-        this._visibilityExpression = createVisibilityExpression(this.visibility, globalState);
+        this._visibilityExpression = createVisibilityExpression(this.visibility, `layers[${this.id}].layout.visibility`, globalState);
 
         if (layer.type === 'custom') return;
 
@@ -142,7 +142,7 @@ export abstract class StyleLayer extends Evented {
             this.source = layer.source;
             this.sourceLayer = layer['source-layer'];
             this.filter = layer.filter;
-            this._featureFilter = featureFilter(layer.filter, globalState);
+            this._featureFilter = featureFilter(layer.filter, `layers[${this.id}].filter`, globalState);
         }
 
         if (properties.layout) {
@@ -167,7 +167,7 @@ export abstract class StyleLayer extends Evented {
 
     setFilter(filter: FilterSpecification | void) {
         this.filter = filter;
-        this._featureFilter = featureFilter(filter, this._globalState);
+        this._featureFilter = featureFilter(filter, `layers[${this.id}].filter`, this._globalState);
     }
 
     getCrossfadeParameters() {